+++ controllers/front/OauthTraitAdapter.php
@@ -26,6 +26,7 @@
 use Nosto\Mixins\OauthTrait as NostoSDKOauthTrait;
 use Nosto\Request\Http\HttpRequest as NostoSDKHttpRequest;
 use Nosto\Types\Signup\AccountInterface as NostoSDKAccountInterface;
+use Nosto\NostoException;
 
 /**
  * @property NostoTagging $module
@@ -35,8 +36,8 @@ class OauthTraitAdapter
     use NostoSDKOauthTrait;
 
     private $languageId;
-    /** @var string */
-    private $moduleName;
+    /** @var NostoTagging $module */
+    private $module;
 
     /**
      * Handles the redirect from Nosto oauth2 authorization server when an existing account is
@@ -44,12 +45,12 @@ class OauthTraitAdapter
      * "return_url" sent in the first step of the authorization cycle, and requires it to be from
      * the same domain that the account is configured for and only redirects to that domain.
      *
-     * @param string $moduleName module name
+     * @param NostoTagging $module
      * @return void
      */
-    public function initContent($moduleName)
+    public function initContent($module)
     {
-        $this->moduleName = $moduleName;
+        $this->module = $module;
         $this->languageId = (int)Tools::getValue('language_id', NostoHelperContext::getLanguageId());
         self::connect();
     }
@@ -66,7 +67,7 @@ class OauthTraitAdapter
         $oauthTraitAdapter = $this;
         return NostoHelperContext::runInContext(
             function () use ($oauthTraitAdapter) {
-                return NostoOAuth::loadData($oauthTraitAdapter->moduleName);
+                return NostoOAuth::loadData($oauthTraitAdapter->module->name);
             },
             $oauthTraitAdapter->languageId
         );
@@ -77,9 +78,28 @@ class OauthTraitAdapter
      * the current store view (as defined by the parameter.)
      *
      * @param Nosto\Types\Signup\AccountInterface $account the account to save
+     * @throws NostoException
      */
     public function save(NostoSDKAccountInterface $account)
     {
+        $success = true;
+        NostoHelperContext::runInContextForEachLanguageEachShop(function () use ($account, &$success) {
+            if ($success
+                && $account->getName() === NostoHelperConfig::getAccountName()
+                && NostoHelperAccount::existsAndIsConnected()
+            ) {
+                $success = false;
+            }
+        });
+        if (!$success) {
+            throw new NostoException(
+                sprintf(
+                    'This account is already being used by "%s". 
+                                        Please create a new account for each store view',
+                    NostoHelperContext::getShop()->name
+                )
+            );
+        }
         NostoHelperAccount::save($account);
     }
 
@@ -113,7 +133,7 @@ class OauthTraitAdapter
     }
 
     /**
-     * Implemented trait method that is responsible for logging an exception to the Magento error
+     * Implemented trait method that is responsible for logging an exception to the Prestashop error
      * log when an error occurs.
      *
      * @param Exception $e the exception to be logged